refacto admin is_clone filter
